Removed the stack system from Edit and replaced it with a single editor view.
